Changes related to patient history - fully working and updates upon submitting info

+++ src/main/java/com/healthsync/entities/Vitals_Results.java
@@ -2,18 +2,18 @@ package com.healthsync.entities;
 
 public class Vitals_Results {
     private int vitals_results_id; // primary key
-    private double height;
-    private double weight;
+    private int height;
+    private int weight;
     private int systolic_bp;
     private int diastolic_bp;
-    private double resting_pulse;
-    private double temperature;
+    private int resting_pulse;
+    private int temperature;
 
     private String patient_id; // foreign key
 
     // Constructors
 
-    public Vitals_Results(int vitals_results_id, double height, double weight, int systolic_bp, int diastolic_bp, double resting_pulse, double temperature, String patient_id) {
+    public Vitals_Results(int vitals_results_id, int height, int weight, int systolic_bp, int diastolic_bp, int resting_pulse, int temperature, String patient_id) {
         this.vitals_results_id = vitals_results_id;
         this.height = height;
         this.weight = weight;
@@ -40,19 +40,19 @@ public class Vitals_Results {
         this.patient_id = patient_id;
     }
 
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
 
-    public void setHeight(double height) {
+    public void setHeight(int height) {
         this.height = height;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
 
-    public void setWeight(double weight) {
+    public void setWeight(int weight) {
         this.weight = weight;
     }
 
@@ -72,19 +72,19 @@ public class Vitals_Results {
         this.diastolic_bp = diastolic_bp;
     }
 
-    public double getResting_pulse() {
+    public int getResting_pulse() {
         return resting_pulse;
     }
 
-    public void setResting_pulse(double resting_pulse) {
+    public void setResting_pulse(int resting_pulse) {
         this.resting_pulse = resting_pulse;
     }
 
-    public double getTemperature() {
+    public int getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(double temperature) {
+    public void setTemperature(int temperature) {
         this.temperature = temperature;
     }
